Debug

+++ flaskapp/common/models/market_valuation.py
@@ -9,11 +9,11 @@ class MarketValuation(db.Model):
     __tablename__ = "t_market_valuation"
     id = db.Column(db.BIGINT, primary_key=True)
     order_id = db.Column(db.BIGINT)
-    graininess = db.Column(db.INT)
-    dimension = db.Column(db.INT)
-    activity = db.Column(db.INT)
-    scale = db.Column(db.INT)
-    correlation = db.Column(db.INT)
+    graininess = db.Column(db.FLOAT)
+    dimension = db.Column(db.FLOAT)
+    activity = db.Column(db.FLOAT)
+    scale = db.Column(db.FLOAT)
+    correlation = db.Column(db.FLOAT)
     suggestion = db.Column(db.BIGINT)
 
     amount = db.Column(db.FLOAT)
